Update scanning recipe, license headers, and rename XmlInputFactory tests
