+++ src/main/java/org/openrewrite/java/security/XmlParserXXEVulnerability.java
@@ -15,15 +15,14 @@
  */
 package org.openrewrite.java.security;
 
-import org.openrewrite.ExecutionContext;
-import org.openrewrite.ScanningRecipe;
-import org.openrewrite.Tree;
-import org.openrewrite.TreeVisitor;
+import org.openrewrite.*;
 import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.security.xml.ExternalDTDAccumulator;
 import org.openrewrite.java.security.xml.TransformerFactoryFixVisitor;
 import org.openrewrite.java.security.xml.XmlInputFactoryFixVisitor;
-import org.openrewrite.xml.tree.Xml;
+
+import java.util.Arrays;
+import java.util.List;
 
 public class XmlParserXXEVulnerability extends ScanningRecipe<ExternalDTDAccumulator> {
 
@@ -50,12 +49,18 @@ public class XmlParserXXEVulnerability extends ScanningRecipe<ExternalDTDAccumul
     //q: What is Java doc format?
     @Override
     public TreeVisitor<?, ExecutionContext> getVisitor(ExternalDTDAccumulator acc) {
+        List<TreeVisitor<?, ExecutionContext>> xmlVisitors = Arrays.asList(
+                XmlInputFactoryFixVisitor.create(acc),
+                TransformerFactoryFixVisitor.create(acc)
+        );
         return new TreeVisitor<Tree, ExecutionContext>() {
             @Override
             public Tree visit(@Nullable Tree tree, ExecutionContext executionContext) {
-                if (!(tree instanceof Xml)) {
-                    tree = XmlInputFactoryFixVisitor.create(acc).visit(tree, executionContext);
-                    tree = TransformerFactoryFixVisitor.create(acc).visit(tree, executionContext);
+                for (TreeVisitor<?, ExecutionContext> xmlVisitor : xmlVisitors) {
+                    //noinspection DataFlowIssue
+                    if (xmlVisitor.isAcceptable((SourceFile) tree, executionContext)) {
+                        tree = xmlVisitor.visit(tree, executionContext);
+                    }
                 }
                 return tree;
             }
