+++ src/test/java/org/openrewrite/java/security/xml/XmlInputFactoryXXEVulnerabilityTest.java
@@ -13,19 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.security;
+package org.openrewrite.java.security.xml;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.openrewrite.DocumentExample;
+import org.openrewrite.java.security.XmlParserXXEVulnerability;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 
 import static org.openrewrite.java.Assertions.java;
 import static org.openrewrite.xml.Assertions.xml;
 
-class XmlParserXXEVulnerabilityTest implements RewriteTest {
+class XmlInputFactoryXXEVulnerabilityTest implements RewriteTest {
     @Override
     public void defaults(RecipeSpec spec) {
         spec.recipe(new XmlParserXXEVulnerability());
